scp: Fix ascii_isdigit build failure
Move ctype replacements to dbctype.c, include it for scp.

Fixes: 7af125aa4bb8 ("Move ctype replacement functions to dbutil.c")
