+++ src/dbctype.h
@@ -1,0 +1,13 @@
+#ifndef DBCTYPE_H
+#define DBCTYPE_H
+
+/* Sensible ctype alternatives */
+int ascii_isdigit(char c);
+int ascii_isalpha(char c);
+int ascii_isalnum(char c);
+int ascii_isspace(char c);
+int ascii_isprint(char c);
+char ascii_tolower(char c);
+
+#endif
+
